Add input type validation to raven filter
