+++ httomolibgpu/cuda_kernels/raven_filter.cu
@@ -1,9 +1,10 @@
 #include <cupy/complex.cuh>
 
-extern "C" __global__ void 
+template <typename Type>
+__global__ void 
 raven_filter(
-  complex<float> *input,
-  complex<float> *output,
+  complex<Type> *input,
+  complex<Type> *output,
   int width, int images, int height, 
   int u0, int n, int v0) {
 
@@ -17,17 +18,21 @@ raven_filter(
   int centerx = width / 2;
   int centerz = height / 2;
 
-  complex<float> value = input[pz * width * images + py * width + px];
+  long long index = static_cast<long long>(px) + 
+                    width * static_cast<long long>(py) + 
+                    width * images * static_cast<long long>(pz);
+
+  complex<Type> value = input[index];
   if( pz >= (centerz - v0) && pz < (centerz + v0 + 1) ) {
     
     // +1 needed to match with CPU implementation
-    float base = float(px - centerx + 1) / u0;
-    float power = base;
+    Type base = Type(px - centerx + 1) / u0;
+    Type power = base;
     for( int i = 1; i < 2 * n; i++ )
       power *= base;
 
-    float filtered_value = 1.f / (1.f + power);
-    value *= complex<float>(filtered_value, filtered_value);
+    Type filtered_value = 1.f / (1.f + power);
+    value *= complex<Type>(filtered_value, filtered_value);
   }
 
   // ifftshifting positions
@@ -36,5 +41,9 @@ raven_filter(
   int outX = (px + xshift) % width;
   int outZ = (pz + zshift) % height;
 
-  output[outZ * width * images + py * width + outX] = value;
+  long long outIndex = static_cast<long long>(outX) + 
+                       width * static_cast<long long>(py) + 
+                       width * images * static_cast<long long>(outZ);
+
+  output[outIndex] = value;
 }
