+++ tests/test_prep/test_stripe.py
@@ -3,8 +3,7 @@ import cupy as cp
 from cupy.cuda import nvtx
 import numpy as np
 import pytest
-import pyfftw
-import pyfftw.interfaces.numpy_fft as fft
+
 from httomolibgpu.prep.normalize import normalize
 from httomolibgpu.prep.stripe import (
     remove_stripe_based_sorting,
@@ -13,53 +12,7 @@ from httomolibgpu.prep.stripe import (
     raven_filter,
 )
 from numpy.testing import assert_allclose
-
-def raven_filter_cpu(
-        sinogram,
-        uvalue: int = 20,
-        nvalue: int = 4,
-        vvalue: int = 2,
-        pad_y: int = 20,
-        pad_x: int = 20,
-        pad_method: str = "edge"):
-    
-    # Parameters
-    v0 = vvalue
-    n = nvalue
-    u0 = uvalue
-
-    # Make a padded copy
-    sinogram_padded = cp.pad(sinogram, ((pad_y,pad_y), (0, 0), (pad_x,pad_x)), pad_method).get()
-    
-    # Size
-    height, images, width = sinogram_padded.shape
-    
-    # Generate filter function
-    centerx = np.ceil(width / 2.0) - 1.0
-    centery = np.int16(np.ceil(height / 2.0) - 1)
-    row1 = centery - v0
-    row2 = centery + v0 + 1
-    listx = np.arange(width) - centerx
-    filtershape = 1.0 / (1.0 + np.power(listx / u0, 2 * n))
-    filtershapepad2d = np.zeros((row2 - row1, filtershape.size))
-    filtershapepad2d[:] = np.float64(filtershape)
-    filtercomplex = filtershapepad2d + filtershapepad2d * 1j
-    
-    # Generate filter objects
-    a = pyfftw.empty_aligned((height, images, width), dtype='complex128', n=16)
-    b = pyfftw.empty_aligned((height, images, width), dtype='complex128', n=16)
-    c = pyfftw.empty_aligned((height, images, width), dtype='complex128', n=16)
-    d = pyfftw.empty_aligned((height, images, width), dtype='complex128', n=16)
-    fft_object  = pyfftw.FFTW(a, b, axes=(0, 2))
-    ifft_object = pyfftw.FFTW(c, d, axes=(0, 2), direction='FFTW_BACKWARD')
-    
-    sino = fft.fftshift(fft_object(sinogram_padded), axes=(0, 2))
-    for m in range(sino.shape[1]):
-        sino[row1:row2, m] = sino[row1:row2, m] * filtercomplex
-    sino = ifft_object(fft.ifftshift(sino, axes=(0, 2)))
-    sinogram = sino[pad_y:height-pad_y, :, pad_x:width-pad_x]
-
-    return sinogram.real
+from .stripe_cpu_reference import raven_filter_cpu
 
 def test_remove_stripe_ti_on_data(data, flats, darks):
     # --- testing the CuPy implementation from TomoCupy ---#
@@ -119,10 +72,10 @@ def test_stripe_raven_cupy(data, flats, darks):
 
     data = normalize(data, flats, darks, cutoff=10, minus_log=True)
 
-    data_after_raven_gpu = raven_filter(cp.copy(data)).get()
-    data_after_raven_cpu = raven_filter_cpu(cp.copy(data))
+    data_after_raven_gpu = raven_filter(cp.copy(data))
+    data_after_raven_cpu = cp.asarray(raven_filter_cpu(cp.copy(data).get()))
 
-    assert_allclose(data_after_raven_cpu, data_after_raven_gpu, 0, atol=4e-01)
+    cp.testing.assert_allclose(data_after_raven_cpu, data_after_raven_gpu, rtol=0, atol=4e-01)
 
     data = None  #: free up GPU memory
     # make sure the output is float32
@@ -210,11 +163,11 @@ def test_raven_filter_cpu_performance(ensure_clean_memory):
     data = cp.asarray(data_host, dtype=np.float32)
 
     # do a cold run first
-    raven_filter_cpu(cp.copy(data))
+    raven_filter_cpu(cp.copy(data).get())
 
     start = time.perf_counter_ns()
     for _ in range(10):
-        raven_filter_cpu(cp.copy(data))
+        raven_filter_cpu(cp.copy(data).get())
 
     duration_ms = float(time.perf_counter_ns() - start) * 1e-6 / 10
